Apply IE_HITBONUSRIGHT whether dual wielding or not
It's very hard to find example of this bonus, at least in BG2.
The associated opcode is used in this mod item, where it
provides an additional +2 to the main hand.
From memory this was not specific to whether you were dual wielding
or not. This needs to be verified, although the new behaviour
makes sense to me.
https://avenger981.github.io/doc/doc/rr_core.html#Revised_Short_Sword_of_Backstabbing