Removing stat limit for IE_DISABLEDBUTTON
This works by left shifting 1 by the value of the ACT_*
variables defined in PCStatStruct.h and doing bitwise
comparisons, the limit of 255 means it quickly stops
working.

Obviously this method isn't that effective because we
can only go up to the value of a 32-bit unsigned integer,
and some of these values are greater than that, so really
this entire thing probably needs to be rethought, maybe
taken out of stats and made a bitset.