Removing invisibility on TryToTalk
This is how the original engine behaves.
There are probably more places where invisibility
should be removed.

This is specifically a problem for dialogues that
trigger a cutscene and then attempt to continue.
As the BeginDialog function in GSUtils does not
target invisible creatures, it can interrupt the
script that cannot find a target and leave you
stuck in the cutscene

The specific case this was encountered was in the
rr#fdrya.bcs script (rogue rebalancing) when talking
to the dryad queen in the acorns quest while invisible.