Fix unlimited casting of contingencies and select spells
These use the target type that casts instantly and doesn't deplete.
I'm still suspicious of this but presumably it has been tested
to behave like this. Still the actual spells need to deplete, so
do it in the opcodes for now.

for fx_spell_list depletion needs to happen before we do any work
as it resets the SpellInfo we set up