In Actor::ApplyKit only apply for the kit in question
The major thing this fixes is dual classing. Previously,
because of the workarounds for oozemaster-type permanent
effects, every level after activation all the effects
of the original class of a dual class would get applied
on top of the actor without being removed.

What we are doing now is only applying the clab when the
dual class is reactivated, and in fact only applying clabs
in general for the actual class that is being leveled up,
unless the dual class needs to be reactivated