Making rest interrupts work
There are a few problems with this. The first is that
we aren't performing the checks at all at the moment.
(this was fixed by e06d530078e78f83e9b2784e6943f09073b02fa7)

Even if we are, spawn groups don't work as we check
if there is a creature for the spawn group resref

The chance of being interrupted at rest is miniscule,
and it's calculated at the beginning rather than every hour.
So we never get partial rest. Checking every hour increases
the chance quite a bit.

Fixed spell restoration per hour and give a message even on
interrupt so that you know how long has passed.

Finally, preventing partial rest from adding fatigue and removing
some weird issue where it intoxicates you. It tries to decrease
intoxication but goes negative and wraps around as stats are
unsigned ints.