pathfinder optimization / overhaul
stop using theta* as it is very inefficient. instead just build
an a* path and smooth it out afterwards

do a first pass of the pathfinding algorithm without actors blocking,
and use the maximum distance of the path as a way to clamp the second
pass. this is an attempt to prevent the pathfinding algorithm exhausting
when there are a lot of blocking actors