Making mislead just give improved invisibility
How this worked before is that fx_force_visible
(done when you perform an attack) modified the puppet
to no longer grant invisibility. This didn't take into
account an actual invisibility dispel which is a
different effect.

Instead here we are just applying improved invisibility
in the original effect and removing the puppet code.

Permanent invisibility is completely broken so that
is just not happening anymore. Mislead is more of a
distraction spell now.