correct sign of points being read from data files
Internally, we store region max/min as a point and a size,
but in most data files they are actually stored as two points.
chu files appear to differ.

We need to read the second point as a pair of signed words
rather than unsigned (as we would if reading size) because otherwise
you can get overflow. This causes problems on the secret door in
the shadow thief guild area (ar0305) in bg2. The door is supposed
to disappear when opened, but an unsigned overflow means its bounding
box takes up practically the entire area.