set default projectile speed back to 0
This is to line up with the logic in NextTarget where if
Speed is 0, the current projectile position is set to the target
desired behaviour when the "none/default" projectile is used as
we want the spell to hit the target and resolve instantly.

This not working appropriately was observed in the scs script
PPJON.BCS, where dw#clone.spl is cast on every party member and
then after 1 tick, it casts the spell ag#ai.spl to flip their
allegiances. If the projectile is not instant then this logic
doesn't work.