fixes for remove school/remove type effects
Check if the effect is pending removal (already been dispelled in same
tick, say in a sequencer).

Don't remove FX_DURATION_INSTANT_PERMANENT_AFTER_BONUSES.
This is used by UAI and without, breach will dispel it as it is a
combat protection.

Place projectiles in the same position on the map in order of creation,
and make projectiles with speed 0 update this tick.

SCS secret word applies two separate spells, one for all spells and
one for spell shield. Spell shield should block the first and be
removed by the second, but projectile placement meant that both spells
worked in tandem to remove spell shield and then remove the next effect.

Also need to make sure that projectiles that cast spells and hit at the same time
as others apply their effects in the correct order (before others).