simplify fx_move_to_area
The previous logic was to make the actor persistent, set the new map
on the actor and remove it from the current. Then when script runs
on the new map, take it out of persistent actors and move to correct
position.

This was a bit buggy, sometimes the actor didn't appear and if
it did, seemed to be missing some initialization (no action bar,
couldn't direct to do anything)

Don't see any reason why the actor needs to be made persistent,
just doing a straight move between areas