Scriptable::ClearActions - clear current action in actionoverride
Previous behaviour is not to clear the current action if it has the
AF_OVERRIDE flag, but there is also a comment somewhere about how
it should only happen for queued actions.

Without this, cut244b in tob breaks as there are two ActionOverride
calls in sequence with PlayDeadInterruptable, the former with the param
10000 and the latter with param 0. Second action is clearly supposed
to override the first but because they are are both ActionOverride,
this wasn't happening.