Actor: fix critical + special flags conflict in Damage
The IsVariable field on the damage effect is used for both
critical animations (PST) and Special flags (EE)

All non-PST criticals were setting the value of this field to
4 (UI_CRITICAL), which caused all criticals to also heal the attacker.
This didn't actually do anything related to criticals in the damage function
because the critical variable is set to 0 for non-PST games (to avoid
this conflicting usage of the field).

It also looked like PST critical animations were being set up regardless
of whether the attack was a critical or not. This logic has been fixed.

Also altering the call to this function in PSTOpcodes that looks
incorrect.